requirements.txt
streamlit>=1.24.0
pyttsx3>=2.90
pdfplumber>=0.9.0
python-dotenv==1.0.0
tqdm==4.66.1
loguru>=0.7.0
tenacity>=8.2.0
watchdog==3.0.0
pdf2image>=1.16.3
pytesseract>=0.3.10
Pillow==10.1.0
psutil>=5.9.0
